% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_lr.R
\name{cor_lr}
\alias{cor_lr}
\title{Title}
\usage{
cor_lr(data, x, y, method = "pearson", alternative = "two.sided",
  exact = NULL, conf.level = 0.95, continuity = FALSE)
}
\arguments{
\item{data}{}

\item{x}{}

\item{y}{}

\item{method}{}

\item{alternative}{}

\item{exact}{}

\item{conf.level}{}

\item{continuity}{}
}
\description{
Title
}
