% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_lr.R
\name{corplot_lr}
\alias{corplot_lr}
\title{Creates a point plot, with R and p labels}
\usage{
corplot_lr(data, x, y, color = NULL, shape = NULL, plot = TRUE,
  hjust = -0.1, vjust = 2)
}
\arguments{
\item{data}{Input data frame}

\item{x}{column name, unquoted}

\item{y}{column name, unquoted}

\item{color}{column name, unquoted}

\item{shape}{column name, unquoted}

\item{plot}{TRUE ... the function plots directly, FALSE ... the function returns a ggplot object for further processing.}
}
\value{
ggplot object
}
\description{
Creates a point plot, with R and p labels
}
\examples{

lr_corplot(mtcars, "hp", "mpg")

}
