% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule_figures_rdkit.R
\name{draw_mol}
\alias{draw_mol}
\title{Plot 2D structure based on SMILES string}
\usage{
draw_mol(smiles, filepath = NA, plot = TRUE, size = 600)
}
\arguments{
\item{plot}{logical, if TRUE the molecule will be directly plot}

\item{size}{size of the figure, default is 600}

\item{filenpath}{path to file}
}
\value{
a png figure of the molecule, either saved to disk or directly plot
}
\description{
Plot 2D structure based on SMILES string
}
\examples{

#You need to set anaconda environment prior to running RDKit code
reticulate::use_condaenv(condaenv = "my-rdkit-env", conda = "/Users/lrichter/miniconda3/bin/conda")

#draws molecule directly into plot pane
draw_mol("C[C@]1([C@H]2C[C@H]3[C@H](C(=O)C(=C([C@]3(C(=O)C2=C(C4=C1C=CC=C4O)O)O)O)C(=O)N)N(C)C)O")

in order to crop the white space around pictures ImageMagick can be used.
the command in windows is >magick mogrify -trim *.png
}
