% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_plots.R
\name{lr_corplot}
\alias{lr_corplot}
\title{Title}
\usage{
lr_corplot(data, x, y)
}
\arguments{
\item{y}{}
}
\value{
ggplot object
}
\description{
Creates a simple ggplot with R and p labels
}
\examples{

lr_corplot(mtcars, "hp", "mpg")

}
