% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule_figures_rdkit.R
\name{makeFigTable}
\alias{makeFigTable}
\title{Plot a 2D grid of figures}
\usage{
makeFigTable(smiles, labels = NA, filepath = NA, plot = TRUE,
  size = 600)
}
\arguments{
\item{smiles}{atomic character vector of canonical smiles}

\item{labels}{atomic character vector for molecule labels}

\item{filepath}{path to file}

\item{plot}{logical, if TRUE the molecule will be directly plot}

\item{size}{size of the figure, default is 600}
}
\value{
a png figure of the molecule, either saved to disk or directly plot
}
\description{
Plot a 2D grid of figures
}
\examples{

}
