% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_rdkit.R
\name{sim_maccs}
\alias{sim_maccs}
\title{Title}
\usage{
sim_maccs(smiles1, smiles2, labels1 = NA, labels2 = NA,
  data.frame = FALSE)
}
\arguments{
\item{smiles1}{atomic character vector containing the SMILES strings}

\item{smiles2}{atomic character vector containing the SMILES strings}

\item{labels1}{atomic character vector with compound labels to smiles1}

\item{labels2}{atomic character vector with compound labels to smiles2}

\item{data.frame}{if TRUE then the output will be a data.frame otherwise a matrix}
}
\value{
similarity matrix as a data frame or matrix
}
\description{
Title
}
\examples{

#You need to set anaconda environment prior to running RDKit code
reticulate::use_condaenv(condaenv = "my-rdkit-env", conda = "/Users/lrichter/miniconda3/bin/conda")
reticulate::use_condaenv(condaenv = "my-rdkit-env", conda = "/Users/lrichter/miniconda3/bin/conda")
}
